八叉树

+++ src/gophy/src/mathx/constant.go
@@ -6,10 +6,10 @@ const (
 	PI = math.Pi
 )
 
-func RadiusToEuler(val float32) float32 {
+func RadiusToEuler(val float64) float64 {
 	return val / PI * 180
 }
 
-func EulerToRadius(val float32) float32 {
+func EulerToRadius(val float64) float64 {
 	return val / 180 * PI
 }
